feat(bin/bump-formula-crossplane.sh): Handle automatically major/minor upgrades
Previously the script couldn't handle major or minor upgrades. In such
cases, I would need to manually change a variable inside the script and
also manually rename the old latest file.

In this commit the variable gets the value automatically, and added
logic to identify major/minor bumps, so that the formula file gets
automatically renamed and the new one gets properly created.