ci(bump-crossplane,install-test-crossplane): Move installation
Move the installation and testing of a new/changed formula to a separate
gh action. This way the bump-crossplane will open a PR, but the
installation and brew test will run after the PR is sent.

This way we can install/test formulas that are changed outside of the
bump-crossplane workflow